Rewrite in pure node to support Windows.
It adds two dependencies, but I think it's worth making it work for any OS.
