+++ packages/cli/index.js
@@ -3,7 +3,12 @@
 import * as fs from 'fs/promises';
 import * as path from 'path';
 import * as url from 'url';
-import { $, ProcessOutput } from 'zx';
+import download from 'download';
+import which from 'which';
+import util from 'util';
+import { exec, spawn } from 'child_process';
+const execAsync = util.promisify(exec);
+const spawnAsync = util.promisify(spawn);
 
 const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
@@ -23,10 +28,6 @@ const windows = platform === 'windows';
 export const installDirectory = path.join(__dirname, 'bin');
 export const name = windows ? 'zig.exe' : 'zig';
 export const binaryPath = path.join(installDirectory, name);
-
-// TODO: currently hardcoded, find a way to fetch the latest version
-// const version = require('./package.json').version;
-export const version = '0.10.0-dev.1717+91eb1af91';
 export const extension = windows ? 'zip' : 'tar.xz';
 
 export async function install({ force = false } = {}) {
@@ -46,11 +47,14 @@ export async function install({ force = false } = {}) {
   }
 
   await uninstall();
-  await $`mkdir -p ${installDirectory}`;
+  await fs.mkdir(installDirectory, { recursive: true });
+
+  const data = await fs.readFile('package.json', 'utf8');
+  const version = JSON.parse(data).zig_version;
 
   const url = `https://ziglang.org/builds/zig-${platform}-${arch}-${version}.${extension}`;
 
-  await $`curl -fsSL ${url} | tar xJ -C ${installDirectory} --strip-components=1`;
+  await download(url, installDirectory, { strip: 1 });
 }
 
 export async function run(...args) {
@@ -61,16 +65,61 @@ export async function run(...args) {
   }
 
   try {
-    await $`${binaryPath} ${args}`;
+    await spawnAsync(binaryPath, args, { stdio: 'inherit' });
   } catch (error) {
-    if (error instanceof ProcessOutput && error.exitCode) {
-      // console.error({ stderr: error.stderr, stdout: error.stdout });
-      process.exit(error.exitCode);
+    if (error.code) {
+      process.exit(error.code);
     }
     throw error;
   }
 }
 
 export async function uninstall() {
-  await $`rm -rf ${installDirectory}`;
+  await fs.rm(installDirectory, { recursive: true });
+}
+
+export async function checkInstallation() {
+  try {
+    const stats = await fs.lstat(binaryPath);
+    // if the binary is a symlink, it should link to the system-wide zig
+    if (stats.isSymbolicLink()) {
+      console.log('Skipping zig installation, symlink to system zig exists');
+      process.exit(0);
+    }
+    // an empty ./bin/zig file is used as a placeholder for npm/pnpm/yarn to
+    // create the bin symlink, so if the file already exists and has non-zero
+    // size, and is not a symlink, then zig is already installed locally
+    if (stats.size !== 0) {
+      console.log('Skipping zig installation, binary exists');
+      process.exit(0);
+    }
+  } catch {}
+  try {
+    // remove the node_modules/.bin symlinks from the PATH before checking if a
+    // zig is installed system-wide, to avoid hitting the symlink
+    const pathSep = windows ? ';' : ':';
+    const pathDirs = process.env.PATH.split(pathSep);
+    process.env.PATH = pathDirs
+      .filter(dir => !dir.endsWith(path.join('node_modules', '.bin')))
+      .join(':');
+    // check if there's already an installed zig binary
+    const systemZig = await which('zig');
+
+    const zigVersion = await execAsync('zig version');
+    const systemZigVersion = zigVersion.stdout.trim();
+
+    if (systemZig.length !== 0 && systemZigVersion.length !== 0) {
+      console.log(
+        `Skipping zig installation, ${systemZigVersion} already installed in system`,
+      );
+      console.log(`Creating symlink to: ${systemZig}`);
+      await fs.unlink(binaryPath);
+      await fs.link(systemZig, binaryPath);
+      console.log(
+        `Manually run the zig-install script to install it locally`,
+      );
+      process.exit(0);
+    }
+  } catch (error) {}
+  console.log('Did not detect zig in system, will be installed locally');
 }
